Moved some tests to a separate directory
