+++ src/path_cache.rs
@@ -1476,38 +1476,6 @@ impl<'a, N: Neighborhood> NodeInspector<'a, N> {
 #[cfg(test)]
 mod tests {
     use crate::prelude::*;
-
-    #[test]
-    fn new() {
-        let grid = [
-            [0, 2, 0, 0, 0],
-            [0, 2, 2, 2, 2],
-            [0, 1, 0, 0, 0],
-            [0, 1, 0, 2, 0],
-            [0, 0, 0, 2, 0],
-        ];
-        let (width, height) = (grid.len(), grid[0].len());
-        fn cost_fn(grid: &[[usize; 5]; 5]) -> impl '_ + Fn((usize, usize)) -> isize {
-            move |(x, y)| [1, 10, -1][grid[y][x]]
-        }
-        let pathfinding = PathCache::new(
-            (width, height),
-            cost_fn(&grid),
-            ManhattanNeighborhood::new(width, height),
-            PathCacheConfig::with_chunk_size(3),
-        );
-        let start = (0, 0);
-        let goal = (4, 4);
-        let path = pathfinding.find_path(start, goal, cost_fn(&grid));
-        let path = path.unwrap();
-        let points: Vec<(usize, usize)> = path.collect();
-        #[rustfmt::skip]
-        assert_eq!(
-            points,
-            vec![(0, 1), (0, 2), (0, 3), (0, 4), (1, 4), (2, 4), (2, 3), (2, 2), (3, 2), (4, 2), (4, 3), (4, 4)],
-        );
-    }
-
     #[test]
     fn get_chunk_index() {
         let grid = [
@@ -1542,139 +1510,5 @@ mod tests {
 
         let point = (0, 4);
         assert_eq!(pathfinding.get_chunk_index(point), 2);
-    }
-
-    #[test]
-    fn update_path() {
-        let mut grid = [
-            [0, 2, 0, 0, 0],
-            [0, 2, 2, 2, 2],
-            [0, 1, 0, 0, 0],
-            [0, 1, 0, 2, 0],
-            [0, 0, 0, 2, 0],
-        ];
-        let (width, height) = (grid.len(), grid[0].len());
-        fn cost_fn(grid: &[[usize; 5]; 5]) -> impl '_ + Fn((usize, usize)) -> isize {
-            move |(x, y)| [1, 10, -1][grid[y][x]]
-        }
-
-        let mut pathfinding = PathCache::new(
-            (width, height),
-            cost_fn(&grid),
-            MooreNeighborhood::new(width, height),
-            PathCacheConfig::with_chunk_size(3),
-        );
-
-        let start = (0, 0);
-        let goal = (4, 4);
-        let path = pathfinding.find_path(start, goal, cost_fn(&grid));
-        let path = path.unwrap();
-        let points: Vec<(usize, usize)> = path.collect();
-        #[rustfmt::skip]
-        assert_eq!(
-            points,
-            vec![(0, 1), (0, 2), (0, 3), (1, 4), (2, 3), (3, 2), (4, 3), (4, 4)],
-        );
-
-        grid[2][1] = 0;
-        let changed_tiles = [(1, 2)];
-
-        pathfinding.tiles_changed(&changed_tiles, cost_fn(&grid));
-
-        let start = (0, 0);
-        let goal = (4, 4);
-        let path = pathfinding.find_path(start, goal, cost_fn(&grid));
-        let path = path.unwrap();
-        let points: Vec<(usize, usize)> = path.collect();
-        #[rustfmt::skip]
-        assert_eq!(
-            points,
-            vec![(0, 1), (1, 2), (2, 2), (3, 2), (4, 3), (4, 4)],
-        );
-
-        // Add walls along chunk borders
-        let start = (0, 0);
-        let goal = (3, 4);
-        let path = pathfinding.find_path(start, goal, cost_fn(&grid));
-        assert!(path.is_some());
-
-        // Vertical wall
-        let changed_tiles: Vec<_> = (0..grid.len()).map(|y| (2, y)).collect();
-        grid.iter_mut().for_each(|row| row[2] = 2);
-        pathfinding.tiles_changed(&changed_tiles, cost_fn(&grid));
-
-        let path = pathfinding.find_path(start, goal, cost_fn(&grid));
-        assert!(path.is_none());
-
-        let goal = (2, 4);
-        let path = pathfinding.find_path(start, goal, cost_fn(&grid));
-        assert!(path.is_some());
-
-        // Horizontal wall
-        let mut changed_tiles = Vec::new();
-        {
-            let y = 2;
-            for x in 0..grid[y].len() {
-                grid[y][x] = 2;
-                changed_tiles.push((x, y));
-            }
-        }
-        pathfinding.tiles_changed(&changed_tiles, cost_fn(&grid));
-
-        let path = pathfinding.find_path(start, goal, cost_fn(&grid));
-        assert!(path.is_none());
-    }
-
-    #[allow(unused)]
-    // #[test]
-    #[cfg(feature = "parallel")]
-    fn random_test() {
-        use nanorand::Rng;
-        use rayon::prelude::*;
-
-        for &size in [8, 128, 1024].iter() {
-            let mut grid = vec![vec![0; size]; size];
-            grid.par_iter_mut().for_each(|row| {
-                // let mut rng = StdRng::from_entropy();
-                let mut rng = nanorand::tls_rng();
-                row.fill_with(|| rng.generate_range(-2_isize..7));
-            });
-            let cost_fn = |(x, y): (usize, usize)| grid[y][x];
-            for &chunk_size in [8, 16, 64].iter() {
-                println!("size: {}, chunk_size: {}", size, chunk_size);
-                let pathfinding = PathCache::new(
-                    (size, size),
-                    cost_fn,
-                    ManhattanNeighborhood::new(size, size),
-                    PathCacheConfig::with_chunk_size(chunk_size),
-                );
-                // for _ in 0..100 {
-                [0..100].par_iter().for_each(|_| {
-                    let mut rng = nanorand::tls_rng();
-                    let start = (rng.generate_range(0..size), rng.generate_range(0..size));
-                    let goal = (rng.generate_range(0..size), rng.generate_range(0..size));
-                    let a_star_path = pathfinding.grid_a_star(start, goal, cost_fn);
-                    let path = pathfinding.find_path(start, goal, cost_fn);
-                    if a_star_path.is_some() != path.is_some() {
-                        use std::io::Write;
-                        let mut out = std::fs::File::create("cache.txt").unwrap();
-                        writeln!(out, "start: {:?},  goal: {:?}", start, goal).unwrap();
-                        writeln!(out, "a_star_path: {:?}", a_star_path).unwrap();
-                        writeln!(out, "our path: {:?}", path).unwrap();
-                        writeln!(out, "{:#?}", pathfinding).unwrap();
-
-                        let mut out = std::fs::File::create("grid.txt").unwrap();
-                        for row in grid.iter() {
-                            for cell in row.iter() {
-                                write!(out, "{}", if *cell < 0 { '#' } else { ' ' }).unwrap();
-                            }
-                            writeln!(out).unwrap();
-                        }
-                        panic!("Failed");
-                    }
-                });
-                // }
-            }
-        }
     }
 }
